requirements.txt
requests
gradio
json_repair
openai-whisper